+++ internal/app/wwctl/overlay/root.go
@@ -9,10 +9,10 @@ import (
 	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/delete"
 	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/edit"
 	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/imprt"
+	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/info"
 	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/list"
 	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/mkdir"
 	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/show"
-	"github.com/warewulf/warewulf/internal/app/wwctl/overlay/variables"
 )
 
 var (
@@ -36,7 +36,7 @@ func init() {
 	baseCmd.AddCommand(imprt.GetCommand())
 	baseCmd.AddCommand(chmod.GetCommand())
 	baseCmd.AddCommand(chown.GetCommand())
-	baseCmd.AddCommand(variables.GetCommand())
+	baseCmd.AddCommand(info.GetCommand())
 }
 
 // GetRootCommand returns the root cobra.Command for the application.